initial work to add fix for missing older shared folders
