+++ src/persistence/plugin.js
@@ -1,7 +1,7 @@
 import { createIdentifierFromNamespaceDefinition, createNamespace } from './utils';
 import existingNamespaceUpdateInterceptor from './existingNamespaceUpdateInterceptor';
 import MCWSPersistenceProvider from './MCWSPersistenceProvider';
-import missingUserFolderInterceptor from './missingUserFolderInterceptor';
+import oldPersistenceFolderInterceptor from './oldPersistenceFolderInterceptor';
 
 export default function MCWSPersistenceProviderPlugin(configNamespaces) {
     return async function install(openmct) {
@@ -11,8 +11,6 @@ export default function MCWSPersistenceProviderPlugin(configNamespaces) {
         });
         openmct.objects.addRoot(() => rootsPromise);
         const namespaces = configNamespaces.map(createNamespace);
-        let usersNamespace = namespaces.find((namespace) => namespace.containsNamespaces);
-        usersNamespace = structuredClone(usersNamespace);
         const mcwsPersistenceProvider = new MCWSPersistenceProvider(openmct, namespaces);
 
         // install the provider for each persistence space,
@@ -25,11 +23,9 @@ export default function MCWSPersistenceProviderPlugin(configNamespaces) {
         const ROOT_IDENTIFIERS = rootNamespaces.map(createIdentifierFromNamespaceDefinition);
 
         // user namespaces are not required
-        if (usersNamespace) {
-            const checkOldNamespaces = localStorage.getItem(`r5.0_old_namespace_checked:${usersNamespace.key}`) === null;
-            existingNamespaceUpdateInterceptor(openmct, usersNamespace, checkOldNamespaces);
-            missingUserFolderInterceptor(openmct, usersNamespace, ROOT_IDENTIFIERS);
-        }
+        const checkOldNamespaces = localStorage.getItem(`r5.0_old_namespace_checked:${usersNamespace.key}`) === null;
+        existingNamespaceUpdateInterceptor(openmct, usersNamespace, checkOldNamespaces);
+        oldPersistenceFolderInterceptor(openmct, namespaces, ROOT_IDENTIFIERS);
 
         rootsResolve(ROOT_IDENTIFIERS);
     };
